requirement.txt
gradio
langchain-core
langchain-openai
langgraph
python-dotenv
scipy
torch